fix(quic): prevent size_t to int truncation in frame decode functions
Change return type from int to ssize_t for QUIC frame decode functions
to prevent potential truncation when casting size_t consumed bytes.

Changes:
- SocketQUICFrame_decode_crypto: int -> ssize_t
- SocketQUICFrame_decode_stream: int -> ssize_t
- Added sys/types.h to SocketQUICFrame.h for ssize_t definition
- Updated test files to use ssize_t for consumed variable

While QUIC frames are typically small (<1500 bytes), RFC 9000 does not
impose hard limits on reassembled stream size. Using ssize_t (standard
POSIX type for byte counts with error indication) prevents theoretical
truncation for large frames exceeding INT_MAX (2,147,483,647 bytes).

Tests: test_quic_crypto_frame_encode and test_quic_stream_frame_encode
pass with AddressSanitizer and UndefinedBehaviorSanitizer.

Closes #774

+++ include/quic/SocketQUICFrame.h
@@ -9,6 +9,7 @@
 
 #include <stddef.h>
 #include <stdint.h>
+#include <sys/types.h>
 
 #include "core/Arena.h"
 #include "core/Except.h"
@@ -208,8 +209,8 @@ extern int SocketQUICFrame_decode_path_response(const uint8_t *in, size_t len, u
 extern size_t SocketQUICFrame_encode_stream(uint64_t stream_id, uint64_t offset,
                                              const uint8_t *data, size_t len,
                                              int fin, uint8_t *out, size_t out_len);
-extern int SocketQUICFrame_decode_stream(const uint8_t *data, size_t len,
-                                          SocketQUICFrameStream_T *frame);
+extern ssize_t SocketQUICFrame_decode_stream(const uint8_t *data, size_t len,
+                                              SocketQUICFrameStream_T *frame);
 
 /* Stream termination frame encoding (RFC 9000 §19.4-19.5) */
 extern size_t SocketQUICFrame_encode_reset_stream(uint64_t stream_id, uint64_t error_code, uint64_t final_size, uint8_t *out, size_t out_size);
@@ -227,7 +228,7 @@ extern int SocketQUICFrame_decode_new_token(const uint8_t *data, size_t len,
 /* CRYPTO frame encoding/decoding (RFC 9000 §19.6) */
 extern size_t SocketQUICFrame_encode_crypto(uint64_t offset, const uint8_t *data,
                                              size_t len, uint8_t *out, size_t out_len);
-extern int SocketQUICFrame_decode_crypto(const uint8_t *data, size_t len,
-                                          SocketQUICFrameCrypto_T *frame);
+extern ssize_t SocketQUICFrame_decode_crypto(const uint8_t *data, size_t len,
+                                              SocketQUICFrameCrypto_T *frame);
 
 #endif /* SOCKETQUICFRAME_INCLUDED */
